fix(deploys): Fix permissions for deploy endpoint projects
Currently, in order to link specific projects to a deploy, e.g. via the `sentry-cli deploys new` command, users must provide a token with the `project:read` scope. This is inconsistent with the `sentry-cli releases new` command, which allows users to create a new release associated with only some projects by using the `org:read` and `project:release` scopes.

This PR proposes allowing specifying projects for a deploy using a token with `project:releases` scope.

Fixes #78025